% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmodel_subdaily.R,
%   R/rpmodel_subdaily_sun_shade.R
\name{headerControl_dd}
\alias{headerControl_dd}
\title{Auxiliar function for dailydounscaling function that check if the mandatory variables 
exist in the dataset (missing or redundant columns)}
\usage{
headerControl_dd(
  df = dfToCheck,
  colMandatory = listMandatoryToCheck,
  showMsg = F
)

headerControl_dd(
  df = dfToCheck,
  colMandatory = listMandatoryToCheck,
  showMsg = F
)
}
\arguments{
\item{df}{dataframe to check the colnames}

\item{colMandatory}{list of mandatory headers}

\item{showMsg}{Logical. If TRUE, it shows the function messages. Defaults to \code{FALSE}.}
}
\value{
TRUE or FALSE.

TRUE or FALSE.
}
\description{
Auxiliar function for dailydounscaling function that check if the mandatory variables 
exist in the dataset (missing or redundant columns)

Auxiliar function for dailydounscaling function that check if the mandatory variables 
exist in the dataset (missing or redundant columns)
}
