% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance_partition_functions.R
\name{irradiance_partition}
\alias{irradiance_partition}
\title{Calculate irradiance of a sun exposed layer or a shade layer in the context 
of a two-layer big-leaf photosynthesis model. 
The partition is based on de Pury and Farquhar 1997}
\usage{
irradiance_partition(
  sun_shade = c("sun", "shade"),
  TIMESTAMP,
  PPFD,
  LAI,
  PA,
  PA0 = 101325,
  fa = 0.426,
  scattering = 0.15,
  rho_cd = 0.036,
  kd_prime = 0.719,
  lat,
  long
)
}
\arguments{
\item{sun_shade}{character setting the layer to be calculated}

\item{TIMESTAMP}{timestamp in a ymd_hms format (as obtained from the package lubridate)}

\item{PPFD}{numeric vector with incident photosynthetic photon flux density [micromol m-2 s-1]}

\item{LAI}{single value or numeric vector of the same length as PPFD with 
Leaf area index [m2 m-2]}

\item{PA}{single value or numeric vector of the same length as PPFD with 
atmospheric pressure [Pa]}

\item{PA0}{reference atmospheric pressure at sea level [Pa]}

\item{fa}{The proportion of attenuated radiation that reaches the surface as
diffuse radiation. Set to 0.426 after de Pury 1995.}

\item{scattering}{Leaf scattering coefficient of PAR. Set to 0.15 for 
a uniform leaf angle distribution}

\item{rho_cd}{canopy reflection coefficient for diffuse PAR. Set to 0.036.}

\item{kd_prime}{diffuse and scattered diffuse PAR extinction coefficient. 
Set to 0.719}

\item{lat}{latitude in decimal format}

\item{long}{longitude in decimal format}
}
\value{
A numeric vector of sun or shade layer Irradiance [micromol m-2 s-1]
 depending on the variable \code{sun_shade}.
}
\description{
Calculate irradiance of a sun exposed layer or a shade layer in the context 
of a two-layer big-leaf photosynthesis model. 
The partition is based on de Pury and Farquhar 1997
}
\references{
de Pury, D. G. G., and G. D. Farquhar, 1997: Simple scaling of photosynthesis 
 from leaves to canopy without the errors of big-leaf models. 
 Plant Cell Environ, 20 , 537–557.
}
